Function core_arch::x86_64::_mm512_maskz_abs_epi32

pub unsafe fn _mm512_maskz_abs_epi32(k: __mmask16, a: __m512i) -> __m512i

🔬 This is a nightly-only experimental API. (stdsimd)

This is supported on x86-64 and target feature avx512f only.

Computes the absolute value of packed 32-bit integers in a, and store the unsigned results in dst using zeromask k (elements are zeroed out when the corresponding mask bit is not set).

Intel's documentation